feat: 2019 day 12

+++ Helpers/Helpers.cs
@@ -8,16 +8,16 @@ using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace Helpers {
 	public class Helpers {
-		public static int GCF(int a, int b) {
+		public static long GCF(long a, long b) {
 			while (b != 0) {
-				int temp = b;
+				long temp = b;
 				b = a % b;
 				a = temp;
 			}
 			return a;
 		}
 
-		public static int LCM(int a, int b) {
+		public static long LCM(long a, long b) {
 			return (a / GCF(a, b)) * b;
 		}
 	}
